// src/app/shared/store/types/general.type.ts
export enum GeneralInformationActionTypes {
  GET_GENERAL_INFORMATION_OPTIONS = 'GET_GENERAL_INFORMATION_OPTIONS',

  GET_PRODUCT_APPLICATIONS = 'GET_PRODUCT_APPLICATIONS',
  GET_PRODUCT_APPLICATIONS_SUCCESS = 'GET_PRODUCT_APPLICATIONS_SUCCESS',
  GET_PRODUCT_APPLICATIONS_FAILURE = 'GET_PRODUCT_APPLICATIONS_FAILURE',

  UPDATE_GENERAL_INFORMATION_MODEL = 'UPDATE_GENERAL_INFORMATION_MODEL',

  RESET_GENERAL_INFORMATION_CHILDREN = 'RESET_GENERAL_INFORMATION_CHILDREN',

  UPDTAE_PRODUCT_TYPE_FOR_CHILDREN = 'UPDTAE_PRODUCT_TYPE_FOR_CHILDREN',

  SET_PRODUCT_TYPE = 'SET_PRODUCT_TYPE',

  CLEAR_STATE = 'CLEAR_STATE',

  SET_SUBSCRIPTION_STATUS = 'SET_SUBSCRIPTION_STATUS',
  GET_INITIAL_STATE = 'GET_INITIAL_STATE'
}